requirements.txt
# Core Deep Learning Framework
# torch>=2.0.0
# torchvision>=0.15.0
# torchaudio>=2.0.0

# EEG Signal Processing
mne>=1.5.0
scipy>=1.10.0
numpy>=1.24.0

# Data Handling and Visualization
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning Utilities
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0

# Experiment Tracking and Logging
tensorboard>=2.13.0
tqdm>=4.65.0

# Configuration Management
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
joblib>=1.3.0
h5py>=3.9.0

# Optional: For reading EDF files if mne is insufficient
pyedflib>=0.1.30
